requirements.txt
pandas                       
google-cloud-bigquery
pandas_gbq
pytrends
